Create a argp directory, it contains the necessary 	files to build a standalone libargp.a
	* argp: New directory.
	argp-ba.c, argp-eexst.c, argp-fmtstream.c, argp-fmtstream.h
	argp-fs-xinl.c, argp-help.c, argp-namefrob.h, argp-parse.c
	argp-pv.c, argp-pvh.c, argp-xinl.c, argp.h, pin.c.

	* mail/mail.c:  Comment out the code that use readline
	specifics WITH_READLINE.
	* mail/mail.h: Likewised.
	* mail/util.c: Likewised.


And still mail will not compile for QNX ... #&*#$^&#^)(

+++ pop3d/pop3d.c
@@ -233,8 +233,8 @@ pop3_mainloop (int infile, int outfile)
   if (md5shared == NULL)
     pop3_abquit (ERR_NO_MEM);
 
-  snprintf (md5shared, strlen (local_hostname) + 50, "<%d.%ld@%s>", getpid (),
-	    time (NULL), local_hostname);
+  snprintf (md5shared, strlen (local_hostname) + 50, "<%u.%u@%s>", getpid (),
+	    (int)time (NULL), local_hostname);
   free (local_hostname);
 
   fflush (ofile);
